+++ src/fiuba/algo3/tp2/modelo/herramienta/creador/MesaDeTrabajoPicoMetal.java
@@ -15,10 +15,10 @@ public class MesaDeTrabajoPicoMetal extends MesaDeTrabajo{
 		
 		try {
 			agregarMaterial(new UnidadMetal(),new Posicion(new Integer(1), new Integer(1)));
-			agregarMaterial(new UnidadMetal(),new Posicion(new Integer(1), new Integer(2)));
-			agregarMaterial(new UnidadMetal(),new Posicion(new Integer(1), new Integer(3)));
+			agregarMaterial(new UnidadMetal(),new Posicion(new Integer(2), new Integer(1)));
+			agregarMaterial(new UnidadMetal(),new Posicion(new Integer(3), new Integer(1)));
 			agregarMaterial(new UnidadMadera(),new Posicion(new Integer(2), new Integer(2)));
-			agregarMaterial(new UnidadMadera(),new Posicion(new Integer(3), new Integer(2)));
+			agregarMaterial(new UnidadMadera(),new Posicion(new Integer(2), new Integer(3)));
 		} catch(PosicionIncorrectaException e) {
 			throw new NoSePuedeInicializarMesaDeTrabajoException();
 		}
